feat: Makes the App.js the single source of books shelved.
Every other component now get its list of books from the `Apps.js`. Besides that, after moving a book from the `BookSearch` component, the user is not redirected to its shelves anymore. Lastly, it fixes a bug where the book shelf was not getting updated on the `BookSearch`.

This commit closes #7, closes #6, and closes #5.